Add optional ino_size filesystem backend function
This is a port of following commit from upstream yaboot:

	commit f9631a4c18c659a6144a697e0c629fe63a44970f
	Author: Anton Blanchard <anton@samba.org>
	Date:   Thu Jul 8 19:03:35 2010 +0000

Our initrd loader is very fragile and the main reason is that it doesn't
know the size of the initrd. We end up claiming 1MB at a time and failing
completely if the new region isn't contiguous with the previous one.

Now that firmware is often at 32MB (real-base), and kernels have grown
much bigger (CONFIG_FUNCTION_TRACER and CONFIG_RELOCATABLE are two big
reasons), we see this failure a lot.

Create a function ino_size (similar to the silo bootloader) and
implement it for tftp and ext2 backends.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>