fs_swap: Reduce severity of not finding a partition table on boot device.
When booted from a cdrom the fs_swap would exit with FILE_ERR_BADDEV,
this causes fs_open to abort too early.

We don't need to check file->device_kind != FILE_DEVICE_BLOCK as we're
only called when that is true.

Demote the return value to FILE_ERR_BAD_FSYS which allows fs_open to
look for other filesystems on the device.

Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>