fs_of: Increase the LOAD_BUFFER_SIZE to 32MB
Okay this is the last time.  To go bigger than this we'll almost
certainly break smaller systems, if we haven't already.

Also 32MB is as big as std. TFTP will go.  To transfer more than that we
need to support tftp block-size and tftp rollover.  Basically we'd
*require* HPA's tftp daemon and we'd need to write a TFTP client in
yaboot.  So we'll stay at 32MB for a while I think.

Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>