yaboot-devel: ofpath does not generate device name
Hi,

I tried latest version of yaboot on Ubuntu 14.10 on PowerMac G5 Quad.
ofpath script does not generate OF device name for my boot disk.

Additional issue is usage of let statement in the script. This works
with BASH, but default in Ubuntu is DASH.

Both issues fixed in the patch bellow.

best regards,
   Matevz Langus

https://lists.ozlabs.org/pipermail/yaboot-devel/2014-December/000483.html